+++ src/Guide.php
@@ -10,21 +10,24 @@
 
 namespace wbrowar\guide;
 
+use craft\events\DefineFieldLayoutElementsEvent;
 use craft\events\RegisterTemplateRootsEvent;
 use craft\events\RegisterUserPermissionsEvent;
 use craft\fields\data\ColorData;
 use craft\helpers\FileHelper;
 use craft\helpers\UrlHelper;
+use craft\models\FieldLayout;
 use craft\services\UserPermissions;
 use craft\web\View;
 use wbrowar\guide\assetbundles\Guide\GuideAsset;
+use wbrowar\guide\fieldlayoutelements\GuideInclude;
+use wbrowar\guide\models\Settings;
 use wbrowar\guide\services\Guide as GuideService;
 use wbrowar\guide\services\GuideComponents as GuideComponentsService;
 use wbrowar\guide\services\Organizer as OrganizerService;
+use wbrowar\guide\twigextensions\GuideTwigExtension;
 use wbrowar\guide\utilities\ImportExport;
 use wbrowar\guide\variables\GuideVariable;
-use wbrowar\guide\twigextensions\GuideTwigExtension;
-use wbrowar\guide\models\Settings;
 use wbrowar\guide\widgets\Guide as GuideWidget;
 
 use Craft;
@@ -214,7 +217,6 @@ class Guide extends Plugin
             );
         }
 
-
         // Add our utilities
         Event::on(
             Utilities::class,
@@ -239,6 +241,18 @@ class Guide extends Plugin
             }
 
             // Add CP Buttons
+            Craft::$app->view->hook('cp.assets.edit.details', function(&$context) {
+                if ($context['element']->volumeId ?? false) {
+                    // Render sidebar template
+                    return self::$view->renderTemplate('guide/sidebar/sidebar', [
+                        'guides' => self::$plugin->guide->getGuidesForUser(['parentUid' => $context['element']->volume->uid, 'parentType' => 'sidebar']),
+                        'settings' => self::$settings,
+                        'userOperations' => self::$userOperations,
+                    ]);
+                }
+
+                return false;
+            });
             Craft::$app->view->hook('cp.categories.edit.details', function(&$context) {
                 if ($context['category'] ?? false) {
                     // Render sidebar template
@@ -276,11 +290,30 @@ class Guide extends Plugin
                 return false;
             });
 
+            // Add custom field UI elements
+            Event::on(
+                FieldLayout::class,
+                FieldLayout::EVENT_DEFINE_UI_ELEMENTS,
+                function(DefineFieldLayoutElementsEvent $event) {
+                    $event->elements[] = GuideInclude::class;
+                }
+            );
+
             // Add modal template to footer
             Event::on(View::class, View::EVENT_END_BODY, function(Event $event) {
                 if ((Craft::$app ?? false) && (Craft::$app->requestedAction ?? false) && (Craft::$app->requestedAction->id ?? false)) {
-                    if (Craft::$app->requestedAction->id == 'edit-entry' && Craft::$app->controller->actionParams['section'] ?? false) {
-                        $element = Craft::$app->getSections()->getSectionByHandle(Craft::$app->controller->actionParams['section']);
+                    if (Craft::$app->requestedAction->id == 'edit-asset' && Craft::$app->controller->actionParams['assetId'] ?? false) {
+                        $element = Craft::$app->getAssets()->getAssetById(Craft::$app->controller->actionParams['assetId']);
+                        $volume = Craft::$app->getVolumes()->getVolumeById($element->volumeId);
+
+                        if ($element ?? false) {
+                            $guides = self::$plugin->guide->getGuides([
+                                'parentType' => 'sidebar',
+                                'parentUid' => $volume->uid,
+                            ]);
+                        }
+                    } else if (Craft::$app->requestedAction->id == 'edit-category' && Craft::$app->controller->actionParams['groupHandle'] ?? false) {
+                        $element = Craft::$app->getCategories()->getGroupByHandle(Craft::$app->controller->actionParams['groupHandle']);
 
                         if ($element ?? false) {
                             $guides = self::$plugin->guide->getGuides([
@@ -288,8 +321,8 @@ class Guide extends Plugin
                                 'parentUid' => $element->uid,
                             ]);
                         }
-                    } else if (Craft::$app->requestedAction->id == 'edit-category' && Craft::$app->controller->actionParams['groupHandle'] ?? false) {
-                        $element = Craft::$app->getCategories()->getGroupByHandle(Craft::$app->controller->actionParams['groupHandle']);
+                    } else if (Craft::$app->requestedAction->id == 'edit-entry' && Craft::$app->controller->actionParams['section'] ?? false) {
+                        $element = Craft::$app->getSections()->getSectionByHandle(Craft::$app->controller->actionParams['section']);
 
                         if ($element ?? false) {
                             $guides = self::$plugin->guide->getGuides([
@@ -319,6 +352,7 @@ class Guide extends Plugin
             );
         }
 
+        // Display welcome message on install
         Event::on(
             Plugins::class,
             Plugins::EVENT_AFTER_INSTALL_PLUGIN,